Converted FeedInfo Activity to fragment
